Day-16-advance-agent/pyproject.toml
[project]
name = "day-16-advance-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.4.201",
    "openai-agents>=0.0.6",
    "python-dotenv>=1.0.1",
]
